net/api: add "server_socket::is_listening()"
* add "bool server_socket::is_listening()" to tell if the server_socket
  has an associated connection.
* correct the doxygen document of `server_socket::server_socket()`.

without this accessor, we would need to either store server_socket
using a smart pointer or std::optional<> or std::variant<> to tell if
the server_socket instance is associated with a connection or not, or
to use `seastar::do_with()` to initialize it with an instace which is
bound to an address.

with this accessor, we can check if the "server_socket" is listening
or not when, for instance, trying to tear down a server. or just use
it in a `assert()` statement.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20220529105532.2925638-1-tchaikov@gmail.com>