net/api: let "server_socket::local_address()" always return an addr
before this change, `server_socket::local_address()` crashes because
of referencing an uninitialized pointer if this `server_socket` is not
bound to an address. this is not quite developer friendly. moreover,
we could use this member function to tell if the `server_socket` is
listening or not.

after this change, `server_socket::local_address()` returns an
empty address if it is not bound to any address.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20220530135308.1045342-1-tchaikov@gmail.com>