perftune.py: remove duplicates in 'append' parameters when we dump an options file
Parameters that are action='append' are technically lists and a user is allowed to
provide the same value more than once.
However we only care about uniq values in those lists.
We have already taken care of this in the functional part of the script, but we would
still print these duplicates when --dump-options-file is given.

This patch fixes that and leaves only uniq values in the output.

Fixes #1084

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>
Message-Id: <20220601172648.2910946-1-vladz@scylladb.com>