rpc: add an option for an asynchronous connection isolation function
The current RPC API has an option to configure a synchronous function
which determines which scheduling group (isolation configuration) to use
for a connection based on a cookie sent during negotiation.
This API lacks an asynchronous option of doing the same, which means
that no preparations can be made in order to accommodate the isolation.
The most obvious example is that the scheduling group used to isolate
the connection can't be created during handshake, forcing an immediate
rejection of the connection in the application side.
This change adds the possibility for an asynchronous function.
It doesn't break applications that uses the old API.

We also add here unit tests for the new API.
Tests: unit tests (dev, debug)

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>