rpc: Generalize _client_negotiated promise
Client resolves this promise when it finishes the negotiation phase.
In the next patches this promise will be needed for both client and
server, so this patch moves one on class connection.

Other than this, on the connection close the promise is resolved into
exception a bit earlier -- before send loop is stopped. Right now
this changes nothing, next patch depends on it.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>