json: wait for writing final characters of a json document
When [[nodiscard]] was introduced for future<>, the same series
also explicitly marked all cases when the future is not waited-on
with a FIXME. One such fixme exists in JSON utilities and sounds
like an obvious case where we should wait - writing final
characters of a json object (e.g. '}' for an object and ']' for lists).
Message-Id: <8810a17d911ba0e2a129691c2550417fba9c29d4.1654527106.git.sarna@scylladb.com>