reactor: re-raise fatal signals
In 1ef66dedd ("core: Print backtrace on SIGSEGV and SIGABRT"), we
started trapping SIGABRT and SIGSEGV in order to print stack traces.
However, we return from the signal handlers as if nothing happened,
while the user expects the program to terminate (it is dead anyway)
and a core dump to be generated.

For SIGSEGV it will likely work, since the fault will happen again,
but for SIGABRT (or a user-sent SIGSEGV) it will not. This means
SIGABRT can't be used to reliably generate cores.

See scylladb/scylla#9242 for an example.

Fix by re-raising the signal explicitly. With this, pkill -ABRT my_app
will cause a core dump.
Message-Id: <20220613152543.3148811-1-avi@scylladb.com>