compile_commands.json generation in configure
CMake can generate a compile_commands.json file, which is a
semi-standard way of integrating other tools with an existing
build system. The file contains the compiler command line for
every source file, and can be consumed by clangd language
server among other things to learn the true compile flags, which
is helpful for full source browing in an IDE, vim, emacs, etc.