Move function in invoke_on_all
Currently, func is copied in one of the invoke_on_all overloadeds,
but there does not seem to be any reason this should not be
a move instead. The difference is significant when func has captured
a large, moveable object.

See also redpanda-data/redpanda#5137 where an OOM passed
though this unexpected copy.