shared_token_bucket: Handle rate * delta overflows
When a bucket is idle for a long time (1+ hour) the resulting
accumulated tokens render extremely large double value. When converted
to the underlying uint64_t tokens value we get undefined behavior. If it
happens that the integer value is zero (or some other small enough one)
the replenishment stops working.

The fix is in limiting the passed delta with the existing max value for
it so that the calculated tokens fit the underlying integer type.

Also add a g++-only static assertion that all this math works as
expected. Clang obeys standard and thinks std::round() is not constexpr
until c++-23.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>