reactor: don't count make_exception_future etc. in cpp_exceptions metric
The `reactor_cpp_exceptions` metric originally meant how many times
exceptions were thrown on the current shard. Throwing an exception is a
computationally expensive operation, therefore such a metric is helpful
when optimizing the error handling logic.

In f5133d05a9, the code increasing this counter was moved from
_Unwind_RaiseException to log_exception_trace. The latter function is
not only called when an exception is thrown but also in
`make_exception_future` and, in the future, possibly in other places
such as `coroutine::exception`.

This commit brings back the old semantic, i.e. only throws are counted.