metrics: support skip_when_empty
Skip when empty, means that metrics that are not in used, will not be
reported.

A common scenario is that a user register a metric, but that metric is
never used.
The most common case is histogram and summary but it is also valid
for counters.

This patch adds an option to mark a metric with skip_when_empty.
When done so, if a metric was never used (true for histogram, counters
and summaries) it will not be reported.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>