metrics: support aggregate by labels
Aggregate by labels is a mechanism for reporting aggregated results. Most
commonly it is used to report one histogram per node instead of per
shard, but it can be used for counters, and gauges.

This patch adds an option to mark a metric with a vector of labels.
That vector will be part of the metric meta-data so the reporting
layer would be able to aggregate over it.