sstring: fix to_sstring_sprintf() compilation on gcc 12.1.1
The to_sstring_sprintf(T) for one of the supported numeric types T,
uses a temporary stack array of size sizeof(T)*3+2 and sprintf()s
into it.

We *know* that the length "len" returned by sprintf() will be less than
the buffer's size, but the compiler doesn't realize this, and when we
try to copy out the "len" bytes from the buffer, Gcc 12.1.1 started
to warn ("stringop-overread") that len may be higher than the size of
the buffer.

We know that this is a false positive, so in this patch we disable
the warning with a pragma.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>