fair_queue: Reserve priority queue on class registration
The priority queue is used to store pointers to classes that have
requests in and plugged. Pushing the popping up priority class can throw
is underlying container expansion takes place. However, when a new class
gets registered we do know how many of then can be queued and can
reserve the priority queue in advance.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>