lowres_timers: Put timeout to infinity if no timers armed
There's a _lowres_next_timeout time point on reactor that denotes
when the next lowres timer expires. If no timers armed this mark is
set to 0 and this makes several other places explicitly check for it
being such. Some hot path ifs can be omitted of setting this time
point to maximum if no timers pending.

tests: unit(dev)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20220706164627.32580-1-xemul@scylladb.com>