reactor: Check the io-properties being YAML::Map
When parsing io-properties the code assumes that it's strucured with
sections, i.e. it's a map (the only supported section nowadays is the
"disks" one, but this fact has nothing to do with this patch). However,
the parsed file can be anything and populating the real io properties
from this "anything" can have weird side effects.

For example, if the YAML contains a list (a.k.a. sequence) the code
fails with an obscure "YAML::InvalidNode" exception, the user has no
glue to understand it was io-properties that failed.

More nasty behavior is when the thing is a scalar value (a string). In
this case it's effectively ignored and the IO stack is configured with
the single unbound queue.

The latter surprise comes whith a CLI sugar -- there are two options to
configure IO properties, one gets the YAML as CLI argument, another
takes its argument as a file name and gets YAML from there. Oblivious
users can easily mix those two up, type

   my_app --io-properties my-io-properties.yaml

and have some good time debugging why the data from the file had no or
unexpected effect on the system behavior.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20220713120835.10889-1-xemul@scylladb.com>