build: find_package() only if necessary
introduce a macro `seastar_set_dep_args()`, to

* pass additional options to find_package()
* set REQUIRED option if the corresponding option
  is enabled explicitly
* do not find_package() if the corresponding option
  is disabled explicitly

this helps to reduce the noise when running cmake
to generate the building system.

also, instead of checking the <package>_FOUND variable
manually, this change leverages the check in
find_package() to stop cmake when a REQUIRED package is
missing while the corresponding option is enabled.

this change also helps make the connection between
a build dependency and its dependent feature more
explicit.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>