io_queue: Split oversized request on submission
When a request whose length is larger than the limit one is submitted,
its split into several sub-requests, all parts are then submitted in the
"correct" order and all futures are waited upon with when_all().

When it resolves the result is collected to return the length of the
contiguous starting requests completed fully.

In fact, the when_all() is a slight overkill here, it's enough to wait
for the last request to complete, all the previous ones should be
resolved by that time. This improvement will come later.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>