io_queue: Keep the iovec memory on io-desc
There are two kinds of io_request's passed to io_queue -- those with a
buffer and those with vector<iovec>. The latter type needs to keep the
vector itself alive throughout the request execution. Nowadays it's
preserved on the .finally() capture.

Things get simpler in the next patch and one allocation is removed here
if the vector<iovec> is carried and kept on io_desc_read_write down the
stack.

Actually, the whole vector is not needed there, the code just needs to
keep the vector memory alive. However, there are no easy ways to detach
the buffer from under a vector and maintain it with the help of a single
pointer.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>