rpc: Make server::connection::negotiate synchronous
When a server connection run its part of negotiation it steps on a
bunch of of "options", and configuring each is potentially sleeping
operation.  Thus some options produce a future and the whole
negotiation() method returns this future to the caller.

Bad news is that different options overwrite futures created by each
other, so only the last one is waited upon, while all the others are
sent to the background. Fortunatelly, there are only two futurized
options, still none should be done in the background.

fixes: #1154

(indentation is deliberately left broken)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>