rpc: Fix formatting on some fmt lib versions
There's a bug in fmt-8 that prefers implicit type cast to bool over
overloaded operator<< when formatting it with the latter. As a result
the formatted string is always "true"/"false" instead of some sane
value.

Marking the operator bool() explicit solves the issue. Also add a test
to catch any violations earlier.

fixes: #1151

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20220725143243.25582-1-xemul@scylladb.com>