batch_flush: Replace circular buffer with slist
Output stream is added to the _flush_batching circular buffer in its
.flush() method that is marked noexcept. However, the buffer's
emplace_back is throwing if fails to allocate memory.

Streams in the list are processed in FIFO order without the ability
to withdraw it from the middle, singly linked list fits this pattern,
has noexcept push_back() and is just one extra pointer in size (on
the stream side).

fixes: #1164
tests: unit(dev)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20220727141903.14981-1-xemul@scylladb.com>