perftune.py: add VIRTIO fast path IRQs ordering functor
The VIRTIO unlike most other drivers allocates 2 IRQs per queue: one for Rx and one for Tx.
It makes sense to bind both IRQs of the same queue to the same CPU.

IRQs names have the following structure:

virtioXYZ-[input|output].NMK

where XYZ is a device index and NMK - a queue index.

A new functor is going to return the NMK for both input and output IRQs ensuring
these are next to each other in the sorted array of IRQs.

We also pre-sort all IRQs numbers array so that the resulting list of VIRTIO IRQ
is numerically sorted too (since Python's sorting is stable).

As a result, since we distribute IRQs evenly among available "IRQ CPUs", we will
distribute Rx and Tx IRQs as evenly as possible amount IRQ CPUs.

In particular, if there is a single IRQ CPU, an even number of them or
if the number of IRQ CPUs is equal to the number if VIRTIO queues Rx and
Tx IRQs of the same VIRTIO queue will be bound to the same IRQ CPU.

Fixes #1165