Create the seastar_memory logger in all builds
Currently the seastar_memory logger is only created when
SEASTAR_DEFAULT_ALLOCATOR is undefined, i.e., when the seastar
allocator is being used. That makes sense in that this logger is only
ever used to emit diagnostics by the seastar allocator.

However, it means that seastar applications which try to adjust the
seastar logger level on the command line will fail because this logger
does not exist. So you have to handle this command line argument
differnetly in release and debug builds.

A more forgiving approach is to allow setting the level of this logger
regardless of the build type. After all, it is valid to set the level of a
loggeer that happens not to log anything for a particular run. This only
takes this a bit further since the logger will *never* log anything in
debug builds.

To do this, we simply move the logger object out of the conditionally
compiled region.