perftune.py: introduce --irq-core-auto-detection-ratio parameter
This new parameter allows overriding a default IRQ core to total cores
ratio during the auto-detection.

By default, this ratio is 16, and it can not be lower than 4 since we
don't allocate a full IRQ core unless there is more than 4 CPU cores
in the 'cpu_mask'.

As usual, this new parameter can be given both via a command line and
in a YAML configuration file.