log: do not inherit formatter<seastar::log_level> from formatter<string_view>
it turns out, in fmt v7.1.3, formatter<std::string_view>::format() is
not marked as const yet. the change which marked the formatters `const`
was introduced in
https://github.com/fmtlib/fmt/commit/bbd6ed5bc57adba7a3fc11804770e75299673761.

before that we cannot assume the constness of this method. since we
don't use the format specifiers supported by formatter<string_view>,
we cannot benefit from inheriting from formatter<string_view>. so,
in this change, instead of making formatter<seastar::log_level> a
derived class of formatter<string_view>, we just implement the
formatter manually so we can preserve the constness of the format()
method.

this change should address the FTBFS with fmt v7.1.3.

Fixes #1189
Signed-off-by: Kefu Chai <tchaikov@gmail.com>

Closes #1191