iostream: Generalize flush and batched flush
When the output_stream::flush() is called it may chose to postpone the
real flushing until the poller kicks it. In the latter case poller calls
::poll_flush() code which effectively does the same.

This patch generalizes both places. The only not-obvious change is the
/zero_copy_put/_fd.put/ one. It's valid, because it's called in non
batch flush branch and in this case the former just calls the latter.

tests: unit(dev)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20220824153637.2498-1-xemul@scylladb.com>