tls_test: Fix spurious fail in test_x509_client_with_builder_system_trust_multiple (et al)
Fixes #1127

google servers have a (pretty short) timeout between connect and expected first
write. If we are delayed inbetween connect and write above (cert verification, scheduling
solar spots or just time sharing on AWS) we could get a short read here. Just retry.
If we get an actual error, it is either on protocol level (exception) or HTTP error.

This is not anything specific to TLS or anything, just particular web server config.

Closes #1193