reactor: add passthru functionality to io_uring backend
passthru is the functionality that enables users to submit an NVMe
command manually constructed by themselves. passthru is also available
via IOCTL, but it is now proper to be used as performance usage because
of IOCTL's overhead.

ZNS is new type of NVMe SSD that only supports sequential write within a
zone. Therefore NVMe commands, which is executed in out-of-order manner,
requires stricted address sequencing overhead.

ZNS append command is new type of NVMe command that targets to relax the
overhead of deciding the block address to place user data. With ZNS
append command, user only submit command to a zone and ZNS SSD decides
the exact block address of user data by its own placement policy. The
assined block address is returned to user via the completion of NVMe
protocol. However append command is not provided as a user level library
like aio_write(). The only way to use append command is submit it via
manually constructed NVMe command.

io_uring library provides passthru with low overhead. Therefore it is
proper to be used to submitting IO commands like append or write. We aim
to use this passthru functionality to utilize append command on user
level.

By utilizing append command on user level, user can mitigate
synchronization overhead for sequencing address rather than using legacy
write command.

This commit contains the implementation of passthru functionality via io
uring library. And it also extents the returning value of IO of
reactor_backend_uring from size_t to io_result which contains size_t and
ssize_t. The second ssize_t is for assigned address of append command
from ZNS SSD. For the other command like write or read, it is fixed to
-1 for compatibility.

Following commit contains the implementation of append command on
file-related classes.

Unit test is done with the our ZNS SSD and test code. However it is not
proper to be upstreamed because it requires a ZNS block device which
cannot be emulated.

Signed-off-by: Heewon Shin <shw096@snu.ac.kr>
Signed-off-by: JinyongHa <jyha200@gmail.com>