file: implement append command
This commit contains implementation of append command on file-related
classes. Almost changes are just copies of write-related functions
except the returned value. The returned value of append is struct
io_result that contains both the status value and assigned block address.

User should be aware of which the 'pos' input parameter should be a start
address of a zone. If 'pos' is not aligned with a zone's starting
address, it will failed with the errno EINVAL.

Added simple example application that write data via append command and
then read data and verify it. The append command is different from the
legacy write command, where the block address of the data is returned
with the command completion (the second field of the io_result structure).
It can be executed on io_uring_reactor backend by giving the option
"--reactor_backend=io_uring", such as
"{source_path}/demos/passthru_append_demo --dev /dev/nvme0n1
--reactor_backend=io_uring"

Signed-off-by: Heewon Shin <shw096@snu.ac.kr>
Signed-off-by: Jinyong Ha <jyha200@gmail.com>