rpc: server: accept: stop conn after process
Currently, process() erases the shared connection from
server._conns and it is destroyed, without stopping.

Since accpet emplaces the connection in _conns,
let's have it erase it when process is done, and make
to stop it befire it;s destroyed.

To prevent a race with `when_all` is server::stop,
serialize _ss_stopped and conn->stop in server::stop().

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>