on_internal_error: add on_fatal_internal_error
Adds an on_internal_error variant that always aborts
after printing the error message, regardless of the
current setting of abort_on_internal_error.

This can be used to replace assert().

Note that it allows the user to log additional
information on top of what would be have been printed
by the failed assert.

Closes #1095

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>