tls: close: move timeout to shutdown
Closing the in and out streams after the timeout
to do a proper bye handshake is essentially part of the
shutdown procedure, not close - which should simply
wait on async stuff to complete.

This change moves the backgrouns with_timout call, to monitor
the _shutdown_promise to the shutdown() method, and leaves
only a single place where `_shutdown_promise.set_value` is
called, after do_shutdown and wait_for_eof are done
(successful or not).

The timeout fiber simply sets _eof and closes in
in and out streams and that should cause do_shutdown
to fail on pipe_closed and for wait_for_eof to pass,
in case they are hanging.

This prevents use-after-free since close waits on
the _shutdown_promise via wait_for_shutdown(),
keeping the object alive.

With that, shutdown() no longer needs to capture shared_from_this,
as we don't want the session to be destroyed in the background after close()
is resolved.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>