input_stream, output_stream: protect against closing more than once
Since the streams have no abort() method, close() is used
both after the stream is fully consumed, or for aborting
the stream (for example, see tls::session::shutdown).

In the latter case, the input or output streams might be closed
more than once.

Note: a shared_promises is used is used to synchronize with
the actual close work to allow waiting on it multiple times.
We use a shared_promise rather than a shared_future because
the latter might throw when move-assigned.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>