rpc: client: move socket param to connected_socket on succesful connect
The socket passed to the rpc::client needs to
go along with the connected_socket that is
passes to the client's rpc::connection class ancestor.

Note that in client::make_stream_sink we return
make_foreign(static_pointer_cast<rpc::connection>(c))
and when eventually stopped, it will need to close the socket
as part of stopping the rpc::connection.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>