rpc: xshard_connection_ptr require stop before destroy
Do not use a foreign_ptr for xshard_connection_ptr
as it allows destroy on a foreing shard, but rather
add a stop method that stops the connection and destroys
it on the owner shard, and enforce calling it before
the xshard_connection_ptr is destroyed.

With that connection::stream_close no longer stops
the connection, but rather it is closed via
xshard_connection_ptr::stop when the
last reference is dropped.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>