test: alloc_test: Actually prevent dead allocation elimination
Turns out dereferencing the object allocated and only freeing it
after we perform the check wasn't enough to prevent the compiler
from eliminating the allocation with certain optimization params
in the build modes.

Let's apply the old and good volatile keyword which tells the
compiler to not perform any optimization on that object.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>
Message-Id: <20220928143402.33353-1-raphaelsc@scylladb.com>