reactor: inline reactor_backend_uring::{read,write}_some()
this change paves the road to the support of network ops like
send(), recv(), sendmsg(), recvmsg(). as the forthcoming implemenation
of reactor_backend_uring::{read,write}_some() would share the same code
structure with the inlined version. so it would be easier for the review
to understand the change we will have.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>