reactor: pass a plain pointer to io_uring_wait_cqes()
there is no need to have an array of `io_uring_cqe*` on the stack.
as the output parameter of `sqe` will simply point to the start address
of the shared ringbuffer of sq if the `io_uring_wait_cqes()` completes.
we could use it to construct a `std::span<io_uring_cqe>` though.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>