reactor: remove unneeded "protected"
`protected` was added when trying to add an inheritance chain of
reactor_backend_uring classes so we can support multiple versions
of linux kernel with different levels of io_uring support. but now
that we are settled on the minimum required version of linux v5.6.

this is not necessary anymore.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>