dns: Handle TCP mode connect failure
Fixes #1244

Iff a TCP mode ARES socket connection fails, we don't get a socket in the socket struct for the virtual fd.
As the connect function cannot fail at this point (engine::connect completion), we can only set all poll flags and wake up the
select emulation we are running.

Here we expect to detect the error in the upcoming sendv call, we however do not check this field being fully initialized -> crash.

Simple fix,. just verify field, if not inited, return ENOTCONN.