reactor: make reactor_backend_selector formattable
With fmt 9, ostream formatters are no longer defaulted in.
Instead, a class as to explicitly opt-in. Do that for reactor_backend_selector
so that reactor.cc compiles without the std::vector partial formatter
specialization which I intend to remove. Include <fmt/ranges.h> so we
don't depend on that partial specialization.