reactor: open fd without O_NONBLOCK when using io_uring
io_uring returns -EAGAIN when performing I/O on an fd marked with
O_NONBLOCK. it only arms poll handler and retries for pollable fd
without being marked with O_NONBLOCK for blocking I/Os. this is the
expected behavior of traditional I/O POSIX APIs, hence the same
applies to io_uring. but the nice thing of io_uring, in this regard,
is that io_uring retries the I/O after the poll handler gets the
event being waited fired, and all of these happen in background.

to take advantage of this feature, we have to open fd without
O_NONBLOCK flag. as we need to perform the speculated I/O in nonblock
mode, while perform the un-speculated I/O in blocking mode. so far,
io_uring does not allow an op to ignore the O_NONBLOCK flag on a
per-op basis or on a per-ring basis. so we cannot opt out this
flag when performing I/O at the time of writing.

in this change, a new virtual function named
`reactor_backend::do_blocking_io()` is added, so that the io_uring
backend can override it. when an fd on which the I/Os are supposed
to be performed with the speculated/un-speculated combination strategy,
is opened, it should be marked without O_NONBLOCK.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>