reactor: use recv(2) for buffered read
before this change, `posix_data_source_impl::get()` tries to
read up to preconfigured size (512, and up to 8192) of data
from socket marked nonblocking, when performing streamed read.
read(2) is used underneath of the streaming io and reactor.
this works great for linux-aio and epoll backend. because we:

- use nonblocking io in the fast path, while
- use wait-for-event-and-perform io in the slow path, if the
  fast path returns EAGAIN.

when it comes io_uring, we want to perform nonblocking I/O first,
in hope that it can be completed immediately for smaller latency,
and turn to blocking I/O in background only when we cannot do it
inlined. when accessing pollable fd opened without O_NONBLOCK flag,
if a nonblocking I/O cannot be completed immediately, io_uring
just arms a poller handler and retries the I/O when the completion
event is triggered for the poller. so we don't have to poll for
read/write if the fast path fails. to make use of this feature,
we need to open an fd to be accessed with io_uring backend without
`O_NONBLOCK` flag, and perform blocked I/O in the fast path.

but the problem is that, the good old `read()` POSIX API does not
offer us a `flags` parameter, so

- if the file to be accessed is opened with `O_NONBLOCK`, we will
  not be able to read the file using a single op prepared with
  `io_uring_prep_read(3)`. we would have a bunch of read cqes
  returning `-EAGAIN`. or we would have to send an extra poll op
  before submitting the read op. neither of these approaches is
  optimal in comparison to the single blocking read explained
  just now.
- if the file to be accessed is opened without `O_NONBLOCK`, we
  will not be able to do perform the nonblocking `read(2)` call
  in the fast path. another option is to submit sqe and reap cqe
  inlined. but that could force us to
  - call all the completion continuations of the completed events,
    even if we have pending preemption requests.
  - and this might incur unexpected latencies.

so, to alleviate this problem, `recv(2)` is used to implement the
buffered read instead. as it supports `flags` parameter, so we can
do nonblocking I/O in the fast path, and do blocking I/O in the
slow path.

in this change, `recv_some()` method is added to the I/O path from
`posix_data_source_impl::get()` all the way down to
`reactor_backend::recv_some()`. this should pave the road to a
more correct implementation of io_uring backend.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>