build: install liburing.pc using stow
before this change, liburing.pc is not installed if we configure
the building system using `./configure ... --cooking liburing` and
expect that CMake will find liburing. but in fact, CMake will reject
the found liburing and continue without liburing enabled, as the found
liburing package is missing version string.

after this change, we install liburing in its top source directory,
so liburing.pc is also installed in addition to its header files and
compiled library.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>