reactor: read() only when data is ready
so far, `read(pollable_fd_state& fd, void* buffer, size_t len)` is only
used by `fsnotifier` for waiting for events. and the fd listening for
the events is opened in `fsnotifier::impl` in `IN_NONBLOCK` mode. as
this facility not closedly coupled by the reactor, it would be difficult
for us to customize the open flags passed to
`file_desc::inotify_init()`. and since fsnotify is rarely sitting in the
critical (i/o) path. it's not a priority for us to make it efficient
unless we have a real-world use case.

before this change, the io_uring backend just call submit the read op
to the ring without waiting for the readable state, so there is good
chance that the corresponding cqe just returns with a `-EAGAIN` result.
this is not what we expect in the use case of `fsnotify`. on the
contrary, what we want is to *wait* until there are events fired.

after this change, we still do the speculated read before submitting
the read op. but we don't submit the read op until the fd is readable.

this change should address the test failure of `Seastar.unit.fsnotifier`
test on 5.11 linux kernel after making io_uring the default backend.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>