http: mark str with [[maybe_unused]]
not all execution code defined by the "action" statements references
the defined `str` lambda, so compiler might warn at seeing unused
variable. with compiling options of `-Werror,-Wunused-variable`, this
even fails the build.

so, in this change, the `str` lambda is marked with `[[maybe_unused]]`
to silence the warning and to fix the FTBFS.

Fixes #1262
Signed-off-by: Kefu Chai <tchaikov@gmail.com>

Closes #1264