reactor: De-futurize IO queues shares updating
This means de-futurizing two places.

The first is io_queue::update_shares_for_class() which is a
futurize_invoke() over a void code for no real reason.

The second is reactor::update_shares_for_queues() which is not
practically useful, because the lambda is non-sleeping and there's no
"real" parallelizm behind it (also there's typically not much of IO
queues to parallelize their update).

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>