build: do not expose Valgrind in SeastarTargets.cmake
This commit resolves issue that despite that the interface
library is claimed PRIVATE, it is still gets populated to
SeastarTargets.cmake. Valgrind is a special type library
that is not required to be linked.

This commit uses solution provided at cmake thread:
https://gitlab.kitware.com/cmake/cmake/-/issues/15415

This fix will cause Valgrind::valgrind is only exposed when
seastar is being built or when it is being linked by another
target in the same build system.

Closes #1263