Make seastar_test constructor noexcept
This constructor cannot throw except in the case of memory exhaustion
causing a bad_alloc to be thrown. Since it it used as a global static
object clang-tidy would like the constructor to be noexcept, so let
us make it so.

In case that an exception is thrown, the error message is the same
for the case of (a) global objects whose constructor throws and (b)
when a noexcept method does throw, so on the (very unlikely) case of
a throw this has the same behavior.