net/dns: fixed socket concurrent access
The problem with DNS queries is caused by the fact that the UDP socket
is used concurrently i.e. write is scheduled while the other one is
pending.

DNS queries are based on c-ares library. The library is initialized with
a set of functions (callbacks) that are called whenever there is a need
to send, receive, create socket, connect or close.

Fixed processing of send requests leading to assertion being triggered
in reactor.

The fix involves checking if send is already pending and if so returning
an error. The error returned from send callback will instruct c-ares
library to retry the query.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
(cherry picked from commit 74d5e9cc02ce9e75be01e4985df6b28a9bac750e)

Closes scylladb/seastar#3143