Add NOLINT for cert-err58-cpp
Disable clang-tidy check cert-err58-cpp in SEASTAR_TEST_CASE and
related macros. This check complains about a possibly-throwing (i.e.,
not noexcept) constructor for global objects. We create such an object
for each test case. Since the constructor can throw (in the unlikely
event of a memory allocation failure) we simply disbale the warning
for these global objects using NOLINT.

In any case, if these objects throw, the process will be terminated
which is an appropriate action for a memory allocation failure at
startup.