util, core: add spawn_process() helper
So we can spawn a subprocess with Seastar.

as `waitpid()` is not directly associated with a fd, in this change
utilizes `pidfd_open()` to obtain an fd referring the child process,
so we can poll for the readiness of the fd before calling `waitpid()`.
but `pidfd_open()` was introduced in linux 5.3. so as an alternative
on systems without this system call, a periodically `waitpid()` call
is used for waiting the completion of the child process.

In our use case, Seastar is used to implement a distributed service.
Before the daemon boots and joins the cluster, it calls an external
script to determine the host's location in the network topology and
failure domain settings, and attaches these information in its RPC
request for joining the cluster sent to the cluster management nodes.
The script can be customized by user so he/she can define their own
deployment policy, as long as the script is an executable and outputs
in the format that can be parsed by the daemon calling the script.

This is a convenient way allowing user to customize the daemon's
behavior without introducing yet another RPC service or a plugin
system.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>