io_request: mark member variables `const`
* add assignment operator
* add move ctor

instead of using the implicitly generated assignment operator,
implement it using `memcpy()`, so we don't mutate the `const` members
in the ops. the move assignment operator is used by
`circular_buffer::erase()`, which relocate the elements using this
operator. `io_sink::_pending_io` is a
`circular_buffer<pending_io_request>`, and `pending_io_request` derives
from `io_request`. so we have to implement the assignment operator
manually. for the sake of completeness, the move ctor is also
implemented.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>