build: compile dpdk with -fpie (position independent executable)
Clang 15 defaults to -fpie, but gcc 12 does not. When building
Seastar with clang, cooking.sh still uses gcc for dpdk. The
mismatch in pie option breaks the link.

Compile dpdk with -fpie to fix. Note an object compiled with -fpie
will still link into a non-pie executable.

This fixes the build on Fedora 37.