prometheus.cc: Support filter by labels
This patch adds the ability to request a subset of the metrics by label
values.

It compares query parameters with the used labels.  A value is a regular
expression, and a missing label counts as an empty string.  So it is
possible to ask for a nonexisting label.

Examples: To get all metrics with scheduling_group_name equals atexit or
main:

http://localhost:9180/metrics?scheduling_group_name=atexit|main

To get all metrics without a ks label:
http://localhost:9180/metrics?ks=

Signed-off-by: Amnon Heiman <amnon@scylladb.com>