parallel utils: add note about shard-local parallelism
A common misconception among people new to seastar is that anything with
"parallel" or "map_reduce" in its name will automatically schedule the
function on all the shards. To their surprise, none of these methods do
that, unless they were explicitly designed for that, like
smp::invoke_on_all(). Add a note to that effect to all such methods, to
avoid the confusion and the wasted time for seastar newbies.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20221115125053.348423-1-bdenes@scylladb.com>